set a role fact and use it in the template
it is a wrong approach to take different actions depending on OS version.
it is not the OS that supports, or not, control-use-cert but unbound.

unfortunately, as the fact cannot be used in a playbook, it does not
make test cases simpler. unless ansible implements facts that show
_available_ package versions, which is very unlikely due to the
overhead, `ansible_distribution_version` is what we have got.